CsvFileWriter, -main, @After(test)
There is a CsvFileWriter in main package, what can write out the functions value with iteration counters. Some of the optimizers call it in the end of their method. I put the important static things to AlgorithmTest.java, so I could delete unnecessary main.java. In addition I wrote an @After method into AlgorithmTest, to write out to the console the datas and the results of the algorithms in an easier way.